+++ programmes/models/__init__.py
@@ -1,0 +1,2 @@
+from .choices import *
+from .models import *
