+++ templates/conventions/journal.html
@@ -58,7 +58,7 @@
                             </td>
                         </tr>
 
-                        {% for evenement in convention.journal %}
+                        {% for evenement in events %}
                             <tr>
                                 {% if action == 'edit' and selected == evenement %}
                                     <td colspan="4">
